feat: hook to deal with games stored on user account
